feat: :+1: organize file structure as feature-first style
