+++ lib/src/app.dart
@@ -6,11 +6,11 @@ import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 
 // Project imports:
+import 'package:yomutube/src/feature/authentication/presentation/login_view.dart';
+import 'package:yomutube/src/feature/settings/presentation/settings_controller.dart';
+import 'package:yomutube/src/feature/settings/presentation/settings_view.dart';
 import 'package:yomutube/src/sample_feature/sample_item_details_view.dart';
 import 'package:yomutube/src/sample_feature/sample_item_list_view.dart';
-import 'package:yomutube/src/settings/settings_controller.dart';
-import 'package:yomutube/src/settings/settings_view.dart';
-import 'package:yomutube/src/view/login_view.dart';
 
 /// The Widget that configures your application.
 class MyApp extends StatelessWidget {
