+++ lib/src/feature/settings/presentation/settings_controller.dart
@@ -2,7 +2,7 @@
 import 'package:flutter/material.dart';
 
 // Project imports:
-import 'package:yomutube/src/settings/settings_service.dart';
+import 'package:yomutube/src/feature/settings/application/settings_service.dart';
 
 /// A class that many Widgets can interact with to read user settings, update
 /// user settings, or listen to user settings changes.
